Use reconnect option with MQTT Paho
Mobile devices should now reconnect when the browser comes to the
foreground or when the device is unlocked. This means they should update
to the current game state within a few seconds of coming back into view.

Also moved the MQTT Paho library from using a CDN to a static file on
our server, since the newest version isn't available on CDN.