Move hideMessagesBefore filter out of a computed property
Previously, the messages before messagereset would breifly render before 
hideMessagesBefore was updated to the current broker value. This is 
because the computed  property would run first with hideMessagesBefore = 
0. The hideMessagesBefore filter has been removed from that computed 
property (and the property renamed to sortedMessages). Now, on any 
message or messagereset topic update, the new message or existing 
messages list is filtered.

This will indeed have the unfortunate side effect that you can no longer 
"undo" a messagereset send. Once the topic has been updated, all 
messages (not currently on the broker) before that time are deleted 
(rather than hidden).