Add flags tracking panel to judges' page game tab
Nothing is synced over MQTT or to the server database, but the state of 
the tabel is saved in the browser's local storage, so it persists across 
page loads.

This does most of what I want for #14, but I will leave that issue open 
for the possibility of adding MQTT support or a scoring system.